requirements.txt
langchain==1.2.10
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-huggingface==1.2.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langgraph>=0.2.0
chromadb
sentence-transformers
datasets
pandas
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
grandalf